ungoliant/requirements.txt
python-dotenv
langchain-openai==0.3.33
langchain==0.3.15   
langgraph==0.2.67
ipython
pandas_market_calendars
pytest